# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Controlled pairwise video comparison platform for flooding/explosion research.",
        "title": "Video Comparison Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/analytics/overview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get analytics overview",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.AnalyticsOverview"
                        }
                    }
                }
            }
        },
        "/admin/analytics/qc": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get QC report",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.QCReport"
                        }
                    }
                }
            }
        },
        "/admin/analytics/study/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get per-study analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Study ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.StudyAnalytics"
                        }
                    }
                }
            }
        },
        "/admin/analytics/study/{id}/pairs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get per-pair analytics for a study",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Study ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.PairStat"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/assets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List video assets (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 100)",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by title (case-insensitive substring)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/assets/free": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all video assets not linked to any pair, for use in pair builder.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List free (unlinked) video assets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/assets/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Deletes a free video asset. Blocked if still linked to a pair or used in presentations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a video asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.AnalyticsOverview": {
            "type": "object",
            "properties": {
                "candidate_win_rate": {
                    "type": "number"
                },
                "completion_rate": {
                    "type": "number"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.EffectAnalytics"
                    }
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.GroupAnalytics"
                    }
                },
                "tie_rate": {
                    "type": "number"
                },
                "total_participants": {
                    "type": "integer"
                },
                "total_responses": {
                    "type": "integer"
                },
                "total_source_items": {
                    "type": "integer"
                }
            }
        },
...
        "service.QCReport": {
            "type": "object",
            "properties": {
                "fast_response_threshold_ms": {
                    "type": "integer"
                },
                "fast_responses": {
                    "type": "integer"
                },
                "straight_lining_profiles": {
                    "type": "integer"
                },
                "total_responses": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CSRFToken": {
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}